PureFun.Contiguous for small size optimizations

PureFun.Contiguous — Module

PureFun.Contiguous implements small-sized versions of lists, sets, and dictionaries. These structures maintain data in contiguous storage, allowing them to leverage the CPU cache to improve the performance of indexing and iteration.

Chunks: densely packed lists

PureFun.Contiguous.StaticChunk — Type

StaticChunk{N,T}
StaticChunk{N}(iter)

Backed by Static Arrays, StaticChunks implement all list functions but are constrained to a maximum size. Useful in conjunction with PureFun.Chunky.@list, which chains together small chunks to build general list types that benefit from data locality.

Examples

This example builds a list with maximum length 8. Until we hit the maximum length, we can use the StaticChunk like any other list type:

julia> xs = Contiguous.StaticChunk{8}(1:3)
3-element PureFun.Contiguous.StaticChunk{8, Int64}
1
2
3

julia> 41 ⇀ 42 ⇀ xs
5-element PureFun.Contiguous.StaticChunk{8, Int64}
41
42
1
2
3

julia> popfirst(xs)
2-element PureFun.Contiguous.StaticChunk{8, Int64}
2
3

PureFun.Contiguous.VectorChunk — Type

VectorChunk{N,T}
VectorChunk{N}(iter)

Backed by Base.Vector, VectorChunks implement all list functions but are constrained to a maximum size. Useful in conjunction with PureFun.Chunky.@list, which chains together small chunks to build general list types that benefit from data locality.

Examples

This example builds a list with maximum length 128. Until we hit the maximum length, we can use the VectorChunk like any other list type:

julia> using PureFun, PureFun.Contiguous
julia> xs = Contiguous.VectorChunk{128}(1:3)
3-element PureFun.Contiguous.VectorChunk{128, Int64}
1
2
3

julia> (41 ⇀ 42 ⇀ xs) ⧺ xs
8-element PureFun.Contiguous.VectorChunk{128, Int64}
41
42
1
2
3
1
2
...

Bits: a set requiring one integer worth of storage

PureFun.Contiguous.Bits — Type

A sparse set with small integer elements.

Examples

julia> b = PureFun.Contiguous.Bits{UInt16}()
0000000000000000

# the 1s in the bitstring mark the elements that are present
julia> b = reduce(push, [1,9,4,15,15], init=b)
0100000100001001

julia> 15 ∈ b, 4 ∈ b, 2 ∈ b
(true, true, false)

BitMap: a small dictionary

Missing docstring.

Missing docstring for PureFun.Contiguous.BitMap. Check Documenter's build log for details.