Reference

PureFun.PFDict — Type

Abstract supertype for immutable dictionaries.

PureFun.PFHeap — Type

PFHeap{T, O<:Base.Order.Ordering}

Supertype for purely functional heaps (aka priority queues) with elements of type T and ordering O. These datastructures give fast access to the minimum element, where comparisons are based on the supplied ordering.

A PFHeap implements push, minimum, delete_min, and merge.

See also PureFun.Pairing.Heap, PureFun.SkewHeap.Heap, and PureFun.FastMerging.Heap

PureFun.PFList — Type

PFList{T}

Supertype for purely functional lists with elements of type T.

A PFList implements cons (equivalent to pushfirst), head, tail, getindex, setindex, append and reverse.

See also PureFun.Linked.List, PureFun.RandomAccess.List, and PureFun.Catenable.List

PureFun.PFQueue — Type

PFQueue{T}

Supertype for purely functional FIFO queues with elements of type T.

A PFQueue implements snoc (equivalent to push), head, and tail

See also PureFun.Batched.Queue, PureFun.Bootstrapped.Queue, and PureFun.RealTime.Queue

PureFun.append — Function

append(xs, ys)
xs ⧺ ys

Concatenate two PFLists.

julia> l1 = PureFun.Linked.List(1:3);

julia> l2 = PureFun.Linked.List(4:6);

julia> l1 ⧺ l2
1
2
3
4
5
6

PureFun.cons — Function

cons(x, xs::PFList)
pushfirst(xs::PFList, x)
x ⇀ xs

Return the PFList that results from adding x to the front of xs.

PureFun.delete_min — Function

delete_min(xs::PFHeap)

return a new heap that is the result of deleting the minimum element from xs, according to the ordering of xs.

PureFun.halfish — Method

halfish(xs)

Split xs roughly in half, and return the two halves as a tuple (front, back). Also returns the new length of the truncated front half

PureFun.head — Function

head(xs)
first(xs)

Return the first element of a PFList or PFQueue. See also tail

PureFun.snoc — Function

snoc(xs::PFQueue, x)
push(xs::PFQueue, x)

Return the PFQueue that results from adding an element to the rear of xs. "snoc is cons from the right."

PureFun.tail — Function

tail(xs)
popfirst(xs)

Return the collection xs without its first element (without modifying xs).

PureFun.update_at — Method

update_at(f::Function, d::PFDict, key, default)

return a new dictionary that is otherwise the same as d, but sets the value associated with key key by appying function f to the current value

StaticArrays.push — Method

push(xs::PFHeap, x)

Return the PFHeap that results from adding x to the collection.

StaticArrays.push — Method

push(xs::PFQueue, x)

Equivalent to snoc

StaticArrays.pushfirst — Method

cons(x, xs::PFList)
pushfirst(xs::PFList, x)
x ⇀ xs

Return the PFList that results from adding x to the front of xs.